+++ src/main/java/homework1/App.java
@@ -1,0 +1,13 @@
+package homework1;
+
+/**
+ * Hello world!
+ *
+ */
+public class App 
+{
+    public static void main( String[] args )
+    {
+        System.out.println("Hello World");
+    }
+}
