--- src/main/java/org/example/App.java
@@ -1,16 +1,0 @@
-package org.example;
-import com.epam.tat.module4.Calculator;
-import org.junit.jupiter.api.Test;
-import static org.junit.jupiter.api.Assertions.*;
-
-/**
- * Hello world!
- *
- */
-public class App 
-{
-    public static void main( String[] args )
-    {
-        System.out.println( "Hello World!" );
-    }
-}
