+++ src/test/java/homework1/BeforeAllTests.java
@@ -1,4 +1,4 @@
-package org.example;
+package homework1;
 
 import com.epam.tat.module4.Calculator;
 import org.junit.jupiter.api.BeforeAll;
