+++ src/test/java/homework1/MultTest.java
@@ -1,4 +1,4 @@
-package org.example;
+package homework1;
 
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.params.ParameterizedTest;
